kernel style ids

+++ include/isoengine/isoengine.h
@@ -437,9 +437,9 @@ bool isoengine_render_draw(void *const engine);
  * @param engine Pointer to the ISO engine instance.
  * @param coords Pointer to the 2D coordinates structure for the object placement.
  * 
- * @return uint64_t The unique identifier of the created 2D object, or 0 if creation failed.
+ * @return uint32_t The unique identifier of the created 2D object, or 0 if creation failed.
  */
-uint64_t isoengine_object2d_create(void *engine, const isoengine_2dcoords *const coords);
+uint32_t isoengine_object2d_create(void *engine, const isoengine_2dcoords *const coords);
 
 /**
  * Registers a keypress callback function for a 2D object in the ISO engine.
@@ -450,7 +450,7 @@ uint64_t isoengine_object2d_create(void *engine, const isoengine_2dcoords *const
  *
  * @return true if the callback was successfully registered, false otherwise.
  */
-bool isoengine_object2d_keypress_callback(void *engine, uint64_t objectid, isoengine_object2d_keypress_callback_func callback);
+bool isoengine_object2d_keypress_callback(void *engine, uint32_t objectid, isoengine_object2d_keypress_callback_func callback);
 
 /**
  * @brief Sets a texture for a 2D object in the ISO engine.
@@ -461,8 +461,8 @@ bool isoengine_object2d_keypress_callback(void *engine, uint64_t objectid, isoen
  * 
  * @return true if the texture was successfully applied, false otherwise.
  */
-bool isoengine_object2d_texture(void *engine, uint64_t objectid, const char *filepng);
+bool isoengine_object2d_texture(void *engine, uint32_t objectid, const char *filepng);
 
-bool isoengine_object2d_delete(void *engine, uint64_t objectid);
+bool isoengine_object2d_delete(void *engine, uint32_t objectid);
 
 #endif
